test: #12 게시글 컨트롤러 테스트 리펙토링
- ArticleControllerTest -> PostControllerTest로 네이밍 변경
- AssertJ를 적용
